Designed logic for calculator
Removed original functions for adding, subtracting, etc.

Operate function serves purpose for calculating two numbers according to operator.

Each button on calculator has appointed value.

Event listener looks for button value and puts target value in the calculate function based on conditions.

Calculate function checks for operator, if found and not "=", it is stored into opValue.

If opValue has not been stored yet, the first value (a) is stored until operator is found.

If the event target is not an operator but first value has been stored, the second value is configured (b).

Each calculation is put as the display value for the user to see.