Created clear function and implemented AC button
When called, the clear function resets calculator by clearing display screen and variables associated with it.

The 'AC' button stands for 'All Clear' which uses the clear function.

The buffer variable checks if a value other than an operator is selected after the operate function.

If so, the calculator is cleared and the display value is the button clicked.